consolidated the startup data in another component
